% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{alert}
\alias{alert}
\title{Alert}
\usage{
alert(width = 12, status = "success", icon = NULL, title = NULL,
  message = NULL, close = FALSE)
}
\arguments{
\item{width}{Largura do alerta.}

\item{status}{O status do alerta, podendo utilizar os valores: danger, info, warning, success.}

\item{icon}{Ícone que será utilizado no título do alerta.}

\item{title}{Título do alerta.}

\item{message}{Mensagem do alerta.}

\item{close}{Um botão para fechar o alerta.}
}
\description{
Uma mensagem de alerta.
}
