% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{appButton}
\alias{appButton}
\title{Button}
\usage{
appButton(id = NULL, label = "label", icon = NULL,
  status = "default", ...)
}
\arguments{
\item{id}{Identificador do botão}

\item{label}{Texto que será exibido no botão}

\item{icon}{icone}

\item{status}{status do botão}
}
\description{
Button
}
