% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{appCircleButton}
\alias{appCircleButton}
\title{appCircleButton}
\usage{
appCircleButton(id = NULL, icon = NULL, status = "default",
  lg = FALSE, ...)
}
\arguments{
\item{id}{id}

\item{icon}{icon}

\item{status}{status}

\item{lg}{lg}

\item{...}{itens}
}
\description{
descricao
}
