% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyDash.R
\name{bodyDash}
\alias{bodyDash}
\title{bodyDash}
\usage{
bodyDash(...)
}
\arguments{
\item{...}{Onde sera passado o \code{\link{tabItem}}.}
}
\description{
O corpo do dashboard, onde serao adicionados os conteudos \code{\link{tabItem}}.
}
