% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box}
\alias{box}
\title{Dashboard box}
\usage{
box(..., width = 12, title = NULL, color = "default",
  message = NULL, minimize = FALSE, close = FALSE)
}
\arguments{
\item{...}{Conteúdo da box.}

\item{width}{Largura da box, podendo passar os valores de 1 à 12}

\item{title}{Título da box}

\item{color}{Cor da borda superior da box, podendo utilizar os valores: primary, success, warning, danger.}

\item{message}{Uma mensagem, que será exibida ao se passar o mouse em cima do ícone.}

\item{minimize}{Podendo ser passados os valores TRUE ou FALSE, onde será adicionado um botão na box, para minimizar a mesma.}

\item{close}{Podendo ser passados os valores TRUE ou FALSE, onde será adicionado um botão na box, para fechar a mesma.}
}
\description{
descricao
}
