% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{callout}
\alias{callout}
\title{Callout}
\usage{
callout(width = 12, status = "success", title = NULL,
  message = NULL)
}
\arguments{
\item{width}{Largura da mensagem.}

\item{status}{O status da mensagem, podendo utilizar os valores: danger, info, warning, success.}

\item{title}{Título da mensagem.}

\item{message}{Mensagem que será exibida, na caixa de mensagem.}
}
\description{
Um caixa com alguma mensagem.
}
