% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\encoding{UTF-8}
\name{carousel}
\alias{carousel}
\title{Carousel}
\usage{
carousel(...)
}
\arguments{
\item{...}{Itens que serão adicionados no carousel}
}
\description{
Função principal do carousel, onde poderá ser adicionado \code{carouselItem}.
}
