% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carouselBlock}
\alias{carouselBlock}
\title{Carousel Block}
\usage{
carouselBlock(..., id = "myCarousel")
}
\arguments{
\item{...}{itens que serão adicionados \code{carouselItem}}

\item{id}{idetificação para o bloco}
}
\description{
descricao
}
\seealso{
\code{\link{carouselItem}}
}
