% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiEelements.R
\name{CarouselItem}
\alias{CarouselItem}
\alias{carouselItem}
\title{Carousel Item}
\usage{
carouselItem(title = NULL, src = NULL)
}
\arguments{
\item{title}{Título que será adicionado em cada slide.}

\item{src}{Caminho da imagem.}
}
\description{
descrição
}
