% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{collapsible}
\alias{collapsible}
\title{Collapsible}
\usage{
collapsible(..., width = 12)
}
\arguments{
\item{...}{Itens que serão adicionados ao collapsible.}

\item{width}{Largura da mensagem.}
}
\description{
Função principal, responsável por agrupar o \code{collapsibleItem}.
}
