% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\encoding{UTF-8}
\name{collapsibleItem}
\alias{collapsibleItem}
\title{Collapsible Item}
\usage{
collapsibleItem(..., id = NULL, title = NULL, icon = NULL,
  color = "default")
}
\arguments{
\item{...}{Conteúdo do collapsibleItem.}

\item{id}{Um identificador para o collapsibleItem, esse identificador deverá ser único e não poderá ter carácter especial.}

\item{title}{Um título para o collapsibleItem.}

\item{icon}{Ícone que será exibido no título.}

\item{color}{Cor da borda superior da box, podendo utilizar os valores: primary, success, warning, danger.}
}
\description{
Box de informação, onde o seu conteúdo poderá ser colapsado.
}
