% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footerDash.R
\name{footerDash}
\alias{footerDash}
\title{Dashboard Footer}
\usage{
footerDash(textLeft = "textLeft", textRight = "textRight")
}
\arguments{
\item{textLeft}{A text footer by left bar}

\item{textRight}{A text footer by right bar#'}
}
\description{
Dashboard Footer
}
