% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footerScriptsDash.R
\name{footerScriptsDash}
\alias{footerScriptsDash}
\title{Dashboard Scripts footer}
\usage{
footerScriptsDash(...)
}
\arguments{
\item{...}{Items to put in the scripts footer by dashboard
.}
}
\description{
Dashboard Scripts footer
}
