% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{infoLink}
\alias{infoLink}
\title{Dashboard infoLink}
\usage{
infoLink(icon = NULL, message = NULL, position = "right", ...)
}
\arguments{
\item{icon}{A icon}

\item{message}{A message}

\item{position}{A position}

\item{...}{Items to put in the dashboard page.}
}
\description{
Dashboard infoLink
}
