% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoBox.R
\name{infoMetric}
\alias{infoMetric}
\title{InfoMetric}
\usage{
infoMetric(width = 12, title = NULL, subTitle = NULL, value = NULL,
  percent = "0\%", colorPercent = "text-navy", icon = NULL, ...)
}
\arguments{
\item{icon}{}
}
\description{
InfoMetric
}
