% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoBox.R
\name{infoProgress}
\alias{infoProgress}
\title{InfoProgress}
\usage{
infoProgress(width = 12, title = NULL, subTitle = NULL,
  value = "33", status = NULL, ...)
}
\arguments{
\item{width}{width}

\item{title}{title}

\item{subTitle}{subTitle}

\item{value}{value}

\item{status}{status}
}
\description{
descricao
}
