% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{label}
\alias{label}
\title{Dashboard label}
\usage{
label(label = "label", status = "success", float = "right", ...)
}
\arguments{
\item{label}{A label.}

\item{status}{A status.}

\item{float}{A float.}
}
\description{
Dashboard label
}
