% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarDash.R
\name{menuDropdown}
\alias{menuDropdown}
\title{Dashboard Menu Dropdown}
\usage{
menuDropdown(text = "Link", icon = NULL, ...)
}
\arguments{
\item{text}{The menu text.}

\item{icon}{The menu icon.}

\item{...}{Items to put in the dropdown menu, using menuItem.}
}
\description{
Dashboard Menu Dropdown
}
\examples{
menuDropdown(
 text = "List",
 menuItem(
   text = "Link1",
   tabName = "link1"
 )
)

}
