% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarDash.R
\name{menuItem}
\alias{menuItem}
\title{Dashboard Menu Item}
\usage{
menuItem(text = "Link", icon = NULL, tabName = NULL, href = NULL)
}
\arguments{
\item{text}{The menu text.}

\item{icon}{The menu icon.}

\item{tabName}{A name the tabItem.}

\item{href}{A link.}
}
\description{
Dashboard Menu Item
}
\examples{
menuItem(
 text = "Home",
 icon = icon("home"),
 tabName = "home",
 href = "http..."
)

}
