% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric}
\alias{metric}
\title{Metric}
\usage{
metric(title = NULL, status = "up", subtitle = NULL)
}
\arguments{
\item{title}{Título para o metric.}

\item{status}{Status do metric, podendo utilizar os valores up e down.}

\item{subtitle}{Subtítulo para o metric.}
}
\description{
Uma caixa de informação, onde será exibido um valor e poderá classificar esse valor como
alto ou baixo.
}
