% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metricBox}
\alias{metricBox}
\title{Metric box}
\usage{
metricBox(title = NULL, value = NULL, percent = NULL,
  subtitle = NULL, status = "success")
}
\arguments{
\item{title}{Título para o metric.}

\item{value}{Valor principal do metric.}

\item{percent}{Valor que será exibido como porcentagem.}

\item{subtitle}{Subtítulo para o metric.}

\item{status}{Status do valor da porcentagem, podendo utilizar os valores success e danger.}
}
\description{
descrição
}
