% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metricProgress}
\alias{metricProgress}
\title{Metric progress}
\usage{
metricProgress(title = NULL, value = 100, subtitle = NULL,
  status = "success")
}
\arguments{
\item{title}{Título para o metric.}

\item{value}{Valor da barra de progresso, podendo utilizar os valores de 0 à 100.}

\item{subtitle}{Subtítulo para o metric.}

\item{status}{Status do progress, podendo utilizar os valores success e danger.}
}
\description{
Uma caixa de informação, onde será exibido um valor, com uma barra de progresso.
}
