% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashNavbar.R
\name{navbarGroup}
\alias{navbarGroup}
\title{Navbar Group}
\usage{
navbarGroup(..., text = NULL)
}
\arguments{
\item{...}{Elementos, como por exemplo. \code{navbarItem}}

\item{text}{Título do grupo ou dropdown.}
}
\description{
Função responsável por agrupar os elementos e assim criar um dropdown.
}
