% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashNavbar.R
\name{navbarLeft}
\alias{navbarLeft}
\title{Navbar Left}
\usage{
navbarLeft(...)
}
\arguments{
\item{...}{Elementos, como por exemplo. \code{navbarItem}}
}
\description{
Função responsável por agrupar os elementos do navbar do lado esquerdo.
}
