% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operPageDash.R
\name{navbarPage}
\alias{navbarPage}
\title{Dashboard Page}
\usage{
navbarPage(title = NULL, navbar = NULL, content = NULL)
}
\arguments{
\item{title}{Um título que será utilizado no display do browser}

\item{navbar}{Um menu Top \code{navbarDash}.}

\item{body}{O conteúdo da página \code{bodyDash}.}

\item{...}{Itens adicionais.}
}
\description{
Dashboard Page
}
\seealso{
\code{\link{sidebarDash}}, \code{\link{navbarDash}}, \code{\link{bodyDash}}.
}
