% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashNavbar.R
\name{navbarRight}
\alias{navbarRight}
\title{Navbar Right}
\usage{
navbarRight(...)
}
\arguments{
\item{...}{Elementos, como por exemplo. \code{navbarItem}}
}
\description{
Função responsável por agrupar os elementos do navbar do lado direito.
}
