% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notification}
\alias{notification}
\title{Dashboard notification}
\usage{
notification(..., icon = NULL, label = NULL, status = "warning")
}
\arguments{
\item{...}{Items to put in the notification item.}

\item{icon}{A icon.}

\item{label}{A label}

\item{status}{A message.}
}
\description{
Dashboard notification
}
\examples{
# Basic dashboard page template
notification(icon = icon("envelope"), label = "1",
notificationItem(
 icon = icon("user", class="fa-3x"),
 title = "novo usu치rio",
 message = "Foi um criado novo usu치rio",
 time = "14:41"),
notificationItem(
 icon = icon("user", class="fa-3x"),
 title = "novo usu치rio",
 message = "Foi um criado novo usu치rio",
 time = "14:41")
)

}
\seealso{
\code{\link{notificationItem}}.
}
