% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notificationItem}
\alias{notificationItem}
\title{Dashboard notification item}
\usage{
notificationItem(icon = NULL, title = NULL, message = NULL,
  time = NULL, ...)
}
\arguments{
\item{icon}{A icon.}

\item{title}{A title.}

\item{message}{A message.}

\item{time}{A time.}

\item{...}{Items to put in the notification item.}
}
\description{
Dashboard notification item
}
