% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashBody.R
\name{operDashBody}
\alias{operDashBody}
\title{Dashboard Body}
\usage{
operDashBody(...)
}
\arguments{
\item{...}{Onde será adicionado o \code{\link{tabItem}}.}
}
\description{
O corpo do dashboard, onde serão adicionados os conteudos \code{\link{tabItem}}.
}
