% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashFooter.R
\name{operDashFooter}
\alias{operDashFooter}
\title{Dashboard Footer}
\usage{
operDashFooter(text_left = NULL, text_right = NULL)
}
\arguments{
\item{text_left}{Texto que ficará localizado do lado esquerdo da página.}

\item{text_right}{Texto que ficará localizado do lado direito da página.}
}
\description{
O rodapé do dashboard.
}
