% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashHeader.R
\name{operDashHeader}
\alias{operDashHeader}
\title{Dashboard Header}
\usage{
operDashHeader(left_menu = NULL, righ_menu = NULL)
}
\arguments{
\item{left_menu}{Itens que serão adicionados a esquerda da barra superior.}

\item{righ_menu}{Itens que serão adicionados a direita da barra superior.}
}
\description{
A função operDashHeader é a função reponsável por adicionar
elementos na barra superior da página.
}
