% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashNavbar.R
\name{operDashNavbar}
\alias{operDashNavbar}
\title{Dashboard Navbar}
\usage{
operDashNavbar(brand = NULL, brand_link = NULL, navbar_left = NULL,
  navbar_right = NULL, color = "#283B4F")
}
\arguments{
\item{color}{}
}
\description{
descricao
}
