% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashPage.R
\name{operDashPage}
\alias{operDashPage}
\title{Dashboard Page}
\usage{
operDashPage(title = NULL, header = NULL, sidebar = NULL,
  body = NULL, footer = NULL)
}
\arguments{
\item{title}{Um título que será utilizado no display do browser}

\item{sidebar}{Um menu lateral \code{sidebar}}

\item{body}{O conteúdo da página \code{body}}

\item{footer}{Script que será adicionado no footer da página}

\item{headScript}{Script que será adicionado no head da página}
}
\description{
descricao
}
\seealso{
\code{\link{sidebar}}, \code{\link{body}}.
}
