% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operPageDash.R
\name{operPage}
\alias{operPage}
\title{Dashboard Page}
\usage{
operPage(title = "", sidebar = NULL, navbar = NULL,
  headScript = NULL, body = NULL, footerScript = NULL)
}
\arguments{
\item{title}{Um título que será utilizado no display do browser}

\item{sidebar}{Um menu lateral \code{sidebarDash}.}

\item{navbar}{Um menu Top \code{navbarDash}.}

\item{headScript}{Script que será adicionado no head da página}

\item{body}{O conteúdo da página \code{bodyDash}.}

\item{footerScript}{Script que será adicionado no footer da página}
}
\description{
Dashboard Page
}
\seealso{
\code{\link{sidebarDash}}, \code{\link{navbarDash}}, \code{\link{bodyDash}}.
}
