% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageDash.R
\name{pageDash}
\alias{pageDash}
\title{Dashboard Page}
\usage{
pageDash(title = "Dashboard Shiny", head = NULL, navbar = NULL,
  body = NULL, footer = NULL, footerScripts = NULL, ...)
}
\arguments{
\item{title}{A title to display in the browser's title bar.}

\item{head}{A head created by \code{headDash}.}

\item{navbar}{A navbar created by \code{navbarDash}.}

\item{body}{A body created by \code{bodyDash}.}

\item{footer}{A footer created by \code{footerDash}.}

\item{footerScripts}{A footerScripts created by \code{footerScriptsDash}.}

\item{...}{Items to put in the dashboard page.}
}
\description{
Dashboard Page
}
\seealso{
\code{\link{headDash}}, \code{\link{navbarDash}},
  \code{\link{bodyDash}}, \code{\link{footerDash}}, \code{\link{footerScriptsDash}}.
}
