% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{panel}
\alias{panel}
\title{Panel}
\usage{
panel(title = NULL, body = NULL, footer = NULL, status = "default",
  width = 12)
}
\arguments{
\item{title}{Um título para o paínel.}

\item{body}{Um conteúdo para o corpo do paínel.}

\item{footer}{Um conteúdo para o rodapé do paínel.}

\item{status}{Um status para o paínel, sendo usado como padrão o valor default,
podendo ser utilizado os valores \strong{default}, \strong{primary}, \strong{success}, \strong{info}, \strong{warning} e \strong{danger}.}

\item{width}{O tamanho do paínel, usando o grid do Bootstrap.
Tendo como padrão o valor 12, podendo ser utilizado os valores de 1 à 12.}
}
\description{
descricao
}
