% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{progressBar}
\alias{progressBar}
\title{progressBar}
\usage{
progressBar(status = "success", animated = FALSE, min = "0",
  value = "45", max = "100", ...)
}
\arguments{
\item{status}{status}

\item{animated}{animated}

\item{min}{min}

\item{value}{value}

\item{max}{max}

\item{...}{itens}
}
\description{
descricao
}
