% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarDash.R
\name{sidebarDash}
\alias{sidebarDash}
\title{Sidebar}
\usage{
sidebarDash(..., imgUser = NULL, userOptions = NULL, brand = NULL)
}
\arguments{
\item{...}{Itens adicionais}

\item{imgUser}{Caminho da imagem}

\item{userOptions}{Box de opções do usuário \code{userOptions}}

\item{brand}{utilizar texto ou imagem}
}
\description{
Sidebar
}
