% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarDash.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Sidebar Item}
\usage{
sidebarItem(text = NULL, icon = NULL, tabName = NULL, ...)
}
\arguments{
\item{text}{text}

\item{icon}{icon}

\item{tabName}{tabName}
}
\description{
descricao
}
