% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashSidebar.R
\name{sidebarMenu}
\alias{sidebarMenu}
\title{Sidebar Menu}
\usage{
sidebarMenu(...)
}
\arguments{
\item{...}{Itens que serão adicionados no menu, como por exemplo \code{sidebarItem}}
}
\description{
A função sidebarMenu será responsável por agrupar os itens do menu.
}
\seealso{
\code{\link{sidebarItem}}.
}
