% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operPageDash.R
\name{sidebarPage}
\alias{sidebarPage}
\title{Dashboard Page}
\usage{
sidebarPage(title = NULL, sidebar = NULL, content = NULL)
}
\arguments{
\item{title}{Um título que será utilizado no display do browser}

\item{sidebar}{Um menu lateral \code{sidebarDash}.}

\item{body}{O conteúdo da página \code{bodyDash}.}

\item{...}{Itens adicionais.}
}
\description{
Dashboard Page
}
\seealso{
\code{\link{sidebarDash}}, \code{\link{navbarDash}}, \code{\link{bodyDash}}.
}
