% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashSidebar.R
\name{sidebarSubItem}
\alias{sidebarSubItem}
\title{Sidebar SubItem}
\usage{
sidebarSubItem(text = NULL, icon = NULL, tabName = NULL,
  href = NULL)
}
\arguments{
\item{text}{text}

\item{icon}{icon}

\item{tabName}{tabName}

\item{href}{href}
}
\description{
descricao
}
