% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashSidebar.R
\name{sidebarUserPanel}
\alias{sidebarUserPanel}
\title{User Panel}
\usage{
sidebarUserPanel(name = NULL, info = NULL, image = NULL)
}
\arguments{
\item{name}{Nome do usuário}

\item{info}{Informações sobre o usuário, como por exemplo o status.}

\item{image}{Endereço da imagem}
}
\description{
A função sidebarUserPanel irá criar um paínel com informações sobre o usuário.
}
\seealso{
\code{\link{operDashSidebar}}.
}
