% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabItem.R
\name{tabItem}
\alias{tabItem}
\title{tabItem}
\usage{
tabItem(tab_name = NULL, title = NULL, ...)
}
\arguments{
\item{tab_name}{Um nome para o tabItem o mesmo nome deverá ser informado no \code{sidebarItem}
ou \code{sidebarSubItem}.}

\item{...}{Conteúdo que será adicionado no corpo da página.}
}
\description{
Função responsável por agrupar conteúdos que serão exibidos no corpo da página.
}
