% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\encoding{UTF-8}
\name{timeline}
\alias{timeline}
\title{Timeline}
\usage{
timeline(..., width = 12)
}
\arguments{
\item{...}{Itens adicionais.}

\item{width}{Largura da timeline}
}
\description{
Função principal da timeline, onde as outras funções \code{timelineLabel} e
\code{timelineItem} poderam ser adicionadas à timeline.
}
