% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiEelements.R
\name{timelineItem}
\alias{timelineItem}
\title{Timeline Item}
\usage{
timelineItem(icon = NULL, icon_color = NULL, time = NULL,
  title = NULL, content = NULL, footer = NULL)
}
\arguments{
\item{icon}{Ícone que será utilizado para marcar o tempo na timeline.}

\item{icon_color}{Cor do ícone, podendo utilizar os valores: blue, red, green, yellow.}

\item{time}{Tempo que será exibido na timeline Item.}

\item{title}{Título da timelineItem.}

\item{content}{Conteúdo da timelineItem.}

\item{footer}{Conteúdo que será adicionado na parte inferior do item da timeline.}
}
\description{
descricao
}
