% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{timelineLabel}
\alias{timelineLabel}
\title{Timeline Label}
\usage{
timelineLabel(color = "green", text = NULL)
}
\arguments{
\item{color}{Cor da label, podendo utilizar os valores: red, green, blue, yellow.}

\item{text}{Texto que será exibido na label.}
}
\description{
Um caixa de texto.
}
