% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{valueBox}
\alias{valueBox}
\title{valueBox}
\usage{
valueBox(width = 12, bg_color = "blue", value = NULL, text = NULL,
  icon = NULL)
}
\arguments{
\item{width}{Largura da box, podendo ser utilizados os valores entre 1 e 12.}

\item{bg_color}{Cor de fundo da box, podendo ser utilizado os valores: blue, green, yellow e red.}

\item{value}{Valor}

\item{text}{Texto}

\item{icon}{Ícone}
}
\description{
Uma caixa de informação.
}
