% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{boxContent}
\alias{boxContent}
\alias{boxTitle}
\title{BoxContent}
\usage{
boxTitle(..., info = NULL, title = "iboxTitle", background = NULL,
  colorText = NULL)

boxContent(...)
}
\arguments{
\item{...}{Itens adicionais.}

\item{info}{info}

\item{title}{title}

\item{background}{background}

\item{colorText}{colorText}

\item{...}{Items to put in the dashboard.}
}
\description{
descricao

descricao
}
