% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{carouselItem}
\alias{carouselItem}
\title{Carousel Item}
\usage{
carouselItem(title = NULL, src = NULL)
}
\arguments{
\item{title}{Título que será adicionado em cada slide.}

\item{src}{Caminho da imagem.}
}
\description{
Item do carousel, podendo passar uma imagem e um título.
}
