% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carouselItem}
\alias{carouselItem}
\title{Carousel Item}
\usage{
carouselItem(imgSrc = NULL, title = NULL, subTitle = NULL)
}
\arguments{
\item{imgSrc}{caminho da imagem}

\item{title}{título para o item do carousel}

\item{subTitle}{subtitulo para o item do carousel}
}
\description{
descricao
}
