% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiEelements.R
\name{collapsibleItem}
\alias{collapsibleItem}
\title{Collapsible Item}
\usage{
collapsibleItem(..., id = NULL, title = NULL, color = "default")
}
\arguments{
\item{...}{Conteúdo do collapsibleItem.}

\item{id}{Um identificador para o collapsibleItem, esse identificador deverá ser único e não poderá ter carácter especial.}

\item{title}{Um título para o collapsibleItem.}

\item{color}{Cor da borda superior da box, podendo utilizar os valores: primary, success, warning, danger.}
}
\description{
descrição
}
