% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footerDash.R
\name{footerDash}
\alias{footerDash}
\title{Dashboard Footer}
\usage{
footerDash(textLeft = "textLeft", textRight = "textRight")
}
\arguments{
\item{textLeft}{Um texto que será adicionado do lado esquerdo da barra}

\item{textRight}{Um texto que será adicionado do lado direito da barra}
}
\description{
Dashboard Footer
}
