% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{infoBox}
\alias{infoBox}
\title{infoBox}
\usage{
infoBox(bg = "aqua", icon = NULL, text = NULL, value = NULL)
}
\arguments{
\item{bg}{Cor de fundo do ícone, podendo ser utilizado os valores: aqua, green, yellow e red.}

\item{icon}{Ícone}

\item{text}{Texto}

\item{value}{Valor}
}
\description{
descrição
}
