% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{infoLink}
\alias{infoLink}
\title{InfoLink}
\usage{
infoLink(icon = NULL, message = NULL, position = "right", ...)
}
\arguments{
\item{icon}{O ícone que será exibido.}

\item{message}{A mensagem que será exibida no balão de informação.}

\item{position}{A posição do balão de informação, podendo utilizar left ou right.}

\item{...}{Itens adicionais.}
}
\description{
InfoLink
}
