% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarDash.R
\name{itemUser}
\alias{itemUser}
\title{User item}
\usage{
itemUser(label = NULL, icon = NULL, tabName = NULL)
}
\arguments{
\item{label}{Texto que será exibido}

\item{icon}{icone}

\item{tabName}{}
}
\description{
User item
}
