% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{label}
\alias{label}
\title{Label}
\usage{
label(text = "label", status = "success", float = "right", ...)
}
\arguments{
\item{text}{O texto que será exibido.}

\item{status}{O status do label que pode ser success, danger, info ou warning.}

\item{float}{A posição do label que pode ser right ou left.}
}
\description{
descricao
}
