% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarDash.R
\name{menuDropdown}
\alias{menuDropdown}
\title{Dashboard Menu Dropdown}
\usage{
menuDropdown(text = "Link", icon = NULL, ...)
}
\arguments{
\item{text}{O texto que será exibido.}

\item{icon}{Um ícone para o item do menu.}

\item{...}{Itens que serão adicionados dentro do menu dropdown, usar o \code{\link{menuItem}}}
}
\description{
Dashboard Menu Dropdown
}
\examples{
menuDropdown(
 text = "List",
 menuItem(
   text = "Link1",
   tabName = "link1"
 )
)

}
