% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarDash.R
\name{menuDropdown}
\alias{menuDropdown}
\title{Menu Dropdown}
\usage{
menuDropdown(text = "Link", icon = NULL, ...)
}
\arguments{
\item{text}{Um texto para o dropdown}

\item{icon}{Um icone para o dropdown}

\item{...}{Itens que serao adicionados no dropdown, use o menuItem}
}
\description{
Menu Dropdown
}
\examples{
menuDropdown(
 text = "List",
 menuItem(
   text = "Link1",
   tabName = "link1"
 )
)

}
