% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarDash.R
\name{menuItem}
\alias{menuItem}
\title{Menu Item}
\usage{
menuItem(text = "Link", icon = NULL, tabName = NULL, href = NULL)
}
\arguments{
\item{text}{O texto que será exibido}

\item{icon}{O icone que será exibido no menuItem}

\item{tabName}{O mesmo nome que será utilizado no tabItem}

\item{href}{Um link}
}
\description{
Menu Item
}
\examples{
menuItem(
 text = "Home",
 icon = icon("home"),
 tabName = "home",
 href = "http..."
)

}
