% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarDash.R
\name{menuItem}
\alias{menuItem}
\title{Dashboard Menu Item}
\usage{
menuItem(text = "Link", icon = NULL, tabName = NULL, href = NULL)
}
\arguments{
\item{text}{O texto que será exibido.}

\item{icon}{Um ícone para o item do menu.}

\item{tabName}{O nome do tabItem que o menuItem está relacionado.}

\item{href}{Um endereço externo.}
}
\description{
Dashboard Menu Item
}
\examples{
menuItem(
 text = "Home",
 icon = icon("home"),
 tabName = "home",
 href = "http..."
)

}
