% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metricBox}
\alias{metricBox}
\title{Metric box}
\usage{
metricBox(title = NULL, value = NULL, percent = NULL,
  subtitle = NULL, status = "success")
}
\arguments{
\item{title}{Título para o metric.}

\item{percent}{}

\item{status}{Status do progress, podendo utilizar os valores success e danger.}
}
\description{
descrição
}
