% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metricProgress}
\alias{metricProgress}
\title{Metric progress}
\usage{
metricProgress(title = NULL, value = 100, subtitle = NULL,
  status = "success")
}
\arguments{
\item{title}{Título para o metric.}

\item{subtitle}{}

\item{status}{Status do progress, podendo utilizar os valores success e danger.}
}
\description{
descrição
}
