% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarDash.R
\name{navbarDash}
\alias{navbarDash}
\title{Dashboard navbarDash}
\usage{
navbarDash(..., brandText = NULL, brandImg = NULL, styleBrand = NULL,
  linkBrand = "#")
}
\arguments{
\item{...}{Items to put in the menu.}

\item{styleBrand}{Customer the CSS brand}

\item{linkBrand}{A link to brand}

\item{brand}{The brand}
}
\description{
Dashboard navbarDash
}
\examples{

navbar = navbarDash(brand = "operDash",
                   menuLeft(
                     menuItem(text = "Home",
                              icon = icon("home"),
                              tabName = "home"),

                     menuDropdown(text = "Lista", icon = icon("list"),
                                  menuItem(
                                    text = "Compras",
                                    tabName = "compras")
                     )
                   )
),
menuRight(
 notification(icon = icon("envelope"), label = "1",
              notificationItem(
                icon = icon("user", class="fa-3x"),
                title = "novo usu치rio",
                message = "Foi um criado novo usu치rio",
                time = "14:41"),
              notificationItem(
                icon = icon("user", class="fa-3x"),
                title = "novo usu치rio",
                message = "Foi um criado novo usu치rio",
                time = "14:41")
 ),
 menuItem(text = "Logout",
          icon = tags$i(class="fas fa-sign-out-alt"),
          tabName = "logout")
)
)

}
\seealso{
\code{\link{menuDropdown}}, \code{\link{menuItem}},
  \code{\link{menuLeft}}, \code{\link{menuRight}}.
}
