% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notification}
\alias{notification}
\title{Dashboard notification}
\usage{
notification(..., icon = NULL, label = NULL, status = "warning")
}
\arguments{
\item{...}{Items to put in the notification item.}

\item{icon}{A icon.}

\item{label}{A label}

\item{status}{A message.}
}
\description{
descricao
}
\seealso{
\code{\link{notificationItem}}.
}
