% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashNavbar.R
\name{operDashNavbar}
\alias{operDashNavbar}
\title{Dashboard Navbar}
\usage{
operDashNavbar(brand = NULL, brand_link = NULL, navbar_left = NULL,
  navbar_right = NULL)
}
\arguments{
\item{...}{dados}
}
\description{
descricao
}
