% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashNavbar.R
\name{operDashNavbar}
\alias{operDashNavbar}
\title{Dashboard Navbar}
\usage{
operDashNavbar(brand = NULL, brand_link = NULL, navbar_left = NULL,
  navbar_right = NULL, color = "#283B4F")
}
\arguments{
\item{brand}{Elemento que será utilizado como marca do site, podendo utilizar imagem ou texto.}

\item{brand_link}{Link ao clicar na marca.}

\item{navbar_left}{Agrupa os elementos do lado esquerdo da navbar. \code{navbarLeft}}

\item{navbar_right}{Agrupa os elementos do lado direito da navbar. \code{navbarRight}}

\item{color}{Cor da navbar, utilizar o valor da cor em hexadecimal.}
}
\description{
Função responsável por criar o menu top.
}
