% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashPage.R
\name{operDashPage}
\alias{operDashPage}
\title{Dashboard Page}
\usage{
operDashPage(title = NULL, header = NULL, sidebar = NULL,
  body = NULL, footer = NULL)
}
\arguments{
\item{title}{Um título que será utilizado no display do browser}

\item{header}{Barra superior da página \code{operDashHeader}}

\item{sidebar}{Menu lateral \code{operDashSidebar}}

\item{body}{O conteúdo da página \code{operDashBody}}

\item{footer}{Barra inferior da página \code{operDashFooter}}
}
\description{
A função operDashPage é a função principal do pacote,
ela engloba todos os elementos que serão adicionados na página.
}
\seealso{
\code{\link{operDashHeader}}, \code{\link{operDashSidebar}},
\code{\link{operDashBody}}, \code{\link{operDashFooter}}.
}
