% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashSidebar.R
\name{operDashSidebar}
\alias{operDashSidebar}
\title{Dashboard Sidebar}
\usage{
operDashSidebar(sidebarUserPanel = NULL, sidebarMenu = NULL)
}
\arguments{
\item{sidebarUserPanel}{Um paínel de informações do usuário conectado no sistema. \code{sidebarUserPanel}}

\item{sidebarMenu}{Responsável por agrupar os itens do menu. \code{sidebarMenu}}
}
\description{
A função operDashSidebar é a função responsável por agrupar os
elementos que ficaram a esquerda da página.
}
\seealso{
\code{\link{sidebarUserPanel}}, \code{\link{sidebarMenu}}.
}
