% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{progressBar}
\alias{progressBar}
\title{Progress Bar}
\usage{
progressBar(width = 12, value = 100, status = "success",
  animated = FALSE)
}
\arguments{
\item{width}{Largura da mensagem.}

\item{value}{O valor da barra de progresso, esse valor deverá ser entre 0 e 100.}

\item{status}{O status da barra de progresso, podendo ser utilizado os valores: primary, success, warning, danger.}

\item{animated}{Caso seja passado o valor TRUE, a barra de progresso ficará animada.}
}
\description{
Uma barra de progresso.
}
