% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarDash.R
\name{sidebarDash}
\alias{sidebarDash}
\title{Sidebar}
\usage{
sidebarDash(..., brand = NULL)
}
\arguments{
\item{...}{Itens adicionais}

\item{brand}{utilizar texto ou imagem}

\item{imgUser}{Caminho da imagem}

\item{userOptions}{Box de opções do usuário \code{userOptions}}
}
\description{
Sidebar
}
