% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashSidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Sidebar Item}
\usage{
sidebarItem(text = NULL, icon = NULL, tabName = NULL, href = NULL,
  sidebarSubItem = NULL)
}
\arguments{
\item{text}{O texto que será exibido}

\item{icon}{Icone que será exibido antes do texto}

\item{tabName}{tabName}

\item{href}{href}

\item{sidebarSubItem}{sidebarSubItem}
}
\description{
descricao
}
