% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operDashSidebar.R
\name{sidebarSubItem}
\alias{sidebarSubItem}
\title{Sidebar SubItem}
\usage{
sidebarSubItem(text = NULL, icon = NULL, tab_name = NULL,
  href = NULL)
}
\arguments{
\item{text}{O texto que será exibido}

\item{icon}{Icone que será exibido antes do texto}

\item{tab_name}{Aqui deverá ser colocado o mesmo nome do tabItem que será
ativado ao clicar nesse item (não podendo utilizar espaço nesse nome)}

\item{href}{Link - Endereço externo}
}
\description{
Essa função será responsável por criar sub menu.
}
