% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timelineBlock}
\alias{timelineBlock}
\alias{timelineItem}
\title{TimelineBlock}
\usage{
timelineBlock(...)

timelineItem(icon = NULL, statusIcon = "primary", title = NULL,
  content = NULL, time = NULL)
}
\arguments{
\item{...}{Itens adicionais.}

\item{time}{}
}
\description{
TimelineBlock

TimelineBlock
}
