% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoBox.R, R/widgets.R
\name{valueBox}
\alias{valueBox}
\title{ValueBox}
\usage{
valueBox(width = 12, bg_color = "aqua", value = NULL, text = NULL,
  icon = NULL)

valueBox(width = 12, bg_color = "aqua", value = NULL, text = NULL,
  icon = NULL)
}
\arguments{
\item{width}{width}

\item{bg_color}{Cor de fundo do ícone, podendo ser utilizado os valores: aqua, green, yellow e red.}

\item{value}{Valor}

\item{text}{Texto}

\item{icon}{icon}

\item{textTop}{textTop}

\item{textSub}{textSub}

\item{status}{status}

\item{...}{itens}

\item{width}{Largura da box, podendo ser utilizados os valores entre 1 e 12.}

\item{icon}{Ícone}
}
\description{
descricao

Uma caixa de informação.
}
