% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoBox.R
\name{valueBox}
\alias{valueBox}
\title{ValueBox}
\usage{
valueBox(width = 12, icon = NULL, textTop = "textTop",
  textSub = "textSub", status = "default", ...)
}
\arguments{
\item{width}{width}

\item{icon}{icon}

\item{textTop}{textTop}

\item{textSub}{textSub}

\item{status}{status}

\item{...}{itens}
}
\description{
descricao
}
